Authenticate requests against role names
Bump version to 1.2.0, to indicate work in R2

+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/controller/AnrXappController.java
@@ -21,6 +21,7 @@ package org.oransc.ric.portal.dashboard.controller;
 
 import java.lang.invoke.MethodHandles;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.oransc.ric.anrxapp.client.api.HealthApi;
@@ -36,6 +37,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.util.Assert;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -64,6 +66,7 @@ public class AnrXappController {
 	public static final String HEALTH_READY_METHOD = "/health/ready";
 	public static final String GNODEBS_METHOD = "/gnodebs";
 	public static final String NCRT_METHOD = "/ncrt";
+	public static final String VERSION_METHOD = DashboardConstants.VERSION_METHOD;
 	// Path parameters
 	public static final String PP_SERVING = "servingcells";
 	public static final String PP_NEIGHBOR = "neighborcells";
@@ -88,8 +91,9 @@ public class AnrXappController {
 	}
 
 	@ApiOperation(value = "Gets the ANR client library MANIFEST.MF property Implementation-Version.", response = SuccessTransport.class)
-	@RequestMapping(value = DashboardConstants.VERSION_METHOD, method = RequestMethod.GET)
+	@RequestMapping(value = VERSION_METHOD, method = RequestMethod.GET)
 	public SuccessTransport getAnrXappClientVersion() {
+		// No role requirement
 		return new SuccessTransport(200, DashboardApplication.getImplementationVersion(HealthApi.class));
 	}
 
@@ -97,6 +101,7 @@ public class AnrXappController {
 	@RequestMapping(value = HEALTH_ALIVE_METHOD, method = RequestMethod.GET)
 	public void getHealthAlive(HttpServletResponse response) {
 		logger.debug("getHealthAlive");
+		// No role requirement
 		healthApi.getHealthAlive();
 		response.setStatus(healthApi.getApiClient().getStatusCode().value());
 	}
@@ -105,25 +110,30 @@ public class AnrXappController {
 	@RequestMapping(value = HEALTH_READY_METHOD, method = RequestMethod.GET)
 	public void getHealthReady(HttpServletResponse response) {
 		logger.debug("getHealthReady");
+		// No role requirement
 		healthApi.getHealthReady();
 		response.setStatus(healthApi.getApiClient().getStatusCode().value());
 	}
 
 	@ApiOperation(value = "Returns list of gNodeB IDs based on NCRT in ANR", response = GgNodeBTable.class)
 	@RequestMapping(value = GNODEBS_METHOD, method = RequestMethod.GET)
-	public GgNodeBTable getGnodebs() {
+	public GgNodeBTable getGnodebs(HttpServletRequest request) {
 		logger.debug("getGnodebs");
+		if (!request.isUserInRole(DashboardConstants.USER_ROLE_UNPRIV))
+			throw new AccessDeniedException("Expected role not found");
 		return ncrtApi.getgNodeB();
 	}
 
 	@ApiOperation(value = "Returns neighbor cell relation table for all gNodeBs or based on query parameters", response = NeighborCellRelationTable.class)
 	@RequestMapping(value = NCRT_METHOD, method = RequestMethod.GET)
-	public NeighborCellRelationTable getNcrt( //
+	public NeighborCellRelationTable getNcrt(HttpServletRequest request, //
 			@RequestParam(name = QP_NODEB, required = false) String ggnbId, //
 			@RequestParam(name = QP_SERVING, required = false) String servingCellNrcgi, //
 			@RequestParam(name = QP_NEIGHBOR, required = false) String neighborCellNrpci) {
 		logger.debug("getNcrt: ggnbid {}, servingCellNrpci {}, neighborCellNrcgi {}", ggnbId, servingCellNrcgi,
 				neighborCellNrpci);
+		if (!request.isUserInRole(DashboardConstants.USER_ROLE_UNPRIV))
+			throw new AccessDeniedException("Expected role not found");
 		return ncrtApi.getNcrt(ggnbId, servingCellNrcgi, neighborCellNrpci);
 	}
 
@@ -131,11 +141,14 @@ public class AnrXappController {
 	@ApiOperation(value = "Modify neighbor cell relation based on Serving Cell NRCGI and Neighbor Cell NRPCI")
 	@RequestMapping(value = NCRT_METHOD + "/" + PP_SERVING + "/{" + PP_SERVING + "}/" + PP_NEIGHBOR + "/{" + PP_NEIGHBOR
 			+ "}", method = RequestMethod.PUT)
-	public void modifyNcrt(@PathVariable(PP_SERVING) String servingCellNrcgi, //
+	public void modifyNcrt(HttpServletRequest request, //
+			@PathVariable(PP_SERVING) String servingCellNrcgi, //
 			@PathVariable(PP_NEIGHBOR) String neighborCellNrpci, //
 			@RequestBody NeighborCellRelationMod ncrMod, HttpServletResponse response) {
 		logger.debug("modifyNcrt: servingCellNrcgi {}, neighborCellNrpci {}, ncrMod {}", servingCellNrcgi,
 				neighborCellNrpci, ncrMod);
+		if (!request.isUserInRole(DashboardConstants.USER_ROLE_PRIV))
+			throw new AccessDeniedException("Expected role not found");
 		ncrtApi.modifyNcrt(servingCellNrcgi, neighborCellNrpci, ncrMod);
 		response.setStatus(healthApi.getApiClient().getStatusCode().value());
 	}
@@ -143,10 +156,13 @@ public class AnrXappController {
 	@ApiOperation(value = "Delete neighbor cell relation based on Serving Cell NRCGI and Neighbor Cell NRPCI")
 	@RequestMapping(value = NCRT_METHOD + "/" + PP_SERVING + "/{" + PP_SERVING + "}/" + PP_NEIGHBOR + "/{" + PP_NEIGHBOR
 			+ "}", method = RequestMethod.DELETE)
-	public void deleteNcrt(@PathVariable(PP_SERVING) String servingCellNrcgi, //
+	public void deleteNcrt(HttpServletRequest request, //
+			@PathVariable(PP_SERVING) String servingCellNrcgi, //
 			@PathVariable(PP_NEIGHBOR) String neighborCellNrpci, //
 			HttpServletResponse response) {
 		logger.debug("deleteNcrt: servingCellNrcgi {}, neighborCellNrpci {}", servingCellNrcgi, neighborCellNrpci);
+		if (!request.isUserInRole(DashboardConstants.USER_ROLE_PRIV))
+			throw new AccessDeniedException("Expected role not found");
 		ncrtApi.deleteNcrt(servingCellNrcgi, neighborCellNrpci);
 		response.setStatus(healthApi.getApiClient().getStatusCode().value());
 	}
